+++ test/hoomd_script/test_ai_pair_dipole.py
@@ -2,6 +2,7 @@
 # Maintainer: mspells
 
 from hoomd_script import *
+context.initialize()
 import unittest
 import os