+++ test/hoomd_script/test_data_access.py
@@ -2,7 +2,7 @@
 # Maintainer: joaander
 
 from hoomd_script import *
-context.initialize();
+context.initialize()
 import unittest
 import os